Multiplikation ist mit double implementiert

+++ Main.java
@@ -20,7 +20,7 @@ public class Main
             summe = summe + i;
         }
         
-        int mult = 1;
+        double mult = 1;
 
         for(int i = 2; i <= obereGrenze; i++)
